Created UI for Create tab
Created components of UI for the Create tab in create_layout.fxml. Added images for the campus map and map pin to resources/drawable. In CreateController.java, implemented two methods, one to handle clicking the create button and the other to handle clicking a location on the campus map. Controller validates study subject, group size, and that a location is set before creating a group. Created a Group class to store a group's information under the Classes package.